# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: My Family Genealogy API
  description: |
    REST API for self-hosted genealogy software. Supports person and family
    management, GEDCOM import/export, pedigree visualization, and search.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API version 1
...
paths:
...
  /browse/surnames:
    get:
      operationId: browseSurnames
      summary: Get surname index with counts
      description: Returns an alphabetical index of surnames with person counts
      tags: [browse]
      parameters:
        - name: letter
          in: query
          description: Filter surnames by starting letter (A-Z)
          schema:
            type: string
            pattern: '^[A-Z]$'
      responses:
        '200':
          description: Surname index or surnames list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurnameIndexResponse'
  /browse/surnames/{surname}/persons:
    parameters:
      - name: surname
        in: path
        required: true
        description: The surname to get people for
        schema:
          type: string
    get:
      operationId: getPersonsBySurname
      summary: Get persons with a specific surname
      tags: [browse]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of persons with the surname
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonList'
  /browse/places:
    get:
      operationId: browsePlaces
      summary: Get place hierarchy with counts
      description: Returns hierarchical place data (Country > State > County > City)
      tags: [browse]
      parameters:
        - name: parent
          in: query
          description: Parent place to get children for (empty for top-level)
          schema:
            type: string
      responses:
        '200':
          description: Place hierarchy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceIndexResponse'
  /browse/places/{place}/persons:
    parameters:
      - name: place
        in: path
        required: true
        description: The place to get people for
        schema:
          type: string
    get:
      operationId: getPersonsByPlace
      summary: Get persons associated with a place
      tags: [browse]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of persons associated with the place
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonList'
...
components:
  parameters:
...
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    ResearchStatus:
      type: string
      description: Confidence level of genealogical data per GPS standards
      enum: [certain, probable, possible, unknown]
      default: unknown
      example: "probable"
    GenDate:
      type: object
      description: Genealogical date with flexible precision
      properties:
        raw:
          type: string
          description: Original date string
          example: "ABT 1850"
        qualifier:
          type: string
          enum: [exact, abt, cal, est, bef, aft, bet, from]
          example: "abt"
        year:
          type: integer
          example: 1850
        month:
          type: integer
          minimum: 1
          maximum: 12
        day:
          type: integer
          minimum: 1
          maximum: 31
        year2:
          type: integer
          description: End year for ranges
        month2:
          type: integer
        day2:
          type: integer
    Person:
      type: object
      required: [id, given_name, surname, version]
      properties:
        id:
          type: string
          format: uuid
        given_name:
          type: string
          maxLength: 100
        surname:
          type: string
          maxLength: 100
        gender:
          type: string
          enum: [male, female, unknown]
        birth_date:
          $ref: '#/components/schemas/GenDate'
        birth_place:
          type: string
        death_date:
          $ref: '#/components/schemas/GenDate'
        death_place:
          type: string
        notes:
          type: string
        research_status:
          $ref: '#/components/schemas/ResearchStatus'
        version:
          type: integer
          format: int64
          description: Optimistic locking version
...
    PersonList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    SurnameIndexResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SurnameEntry'
        total:
          type: integer
          description: Total number of unique surnames
        letter_counts:
          type: array
          items:
            $ref: '#/components/schemas/LetterCount'
          description: Count of surnames by starting letter
    SurnameEntry:
      type: object
      required: [surname, count]
      properties:
        surname:
          type: string
        count:
          type: integer
          description: Number of persons with this surname
    LetterCount:
      type: object
      required: [letter, count]
      properties:
        letter:
          type: string
          description: A-Z letter
        count:
          type: integer
          description: Number of surnames starting with this letter
    PlaceIndexResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlaceEntry'
        total:
          type: integer
          description: Total number of unique places at this level
        breadcrumb:
          type: array
          items:
            type: string
          description: Path from root to current level
    PlaceEntry:
      type: object
      required: [name, count]
      properties:
        name:
          type: string
          description: Place name at this level
        full_name:
          type: string
          description: Full place path
        count:
          type: integer
          description: Number of persons associated with this place
        has_children:
          type: boolean
          description: Whether this place has sub-locations